Add MultiChainFit base, port CmdStanGQ
MultiChainFit holds the shared behavior of fits with one Stan CSV
file per chain: the chains property, list-normalizing from_files
parsing, and cross-chain config validation, with a _comparable_config
hook that subclasses extend with the settings that affect their draw
layout.

CmdStanGQ keeps only its GQ-specific members: previous_fit and the
draw accessors that merge it with the generated quantities. Its
config validation needs were exactly the base ones, so it vanishes
entirely. gq.py shrinks from 815 to 668 lines.
